Remove all Warning

+++ AStarPathfinder/Sources/Render/SFML/Private/SFMLRender.cpp
@@ -22,8 +22,8 @@ namespace
 
 	void DrawQuad2D(FQuad2D& obj)
 	{
-		sf::RectangleShape rectangle(sf::Vector2f(obj.transform.scale.x * CellSize, obj.transform.scale.y * CellSize));
-		rectangle.setPosition(obj.transform.pos.x * CellSize, obj.transform.pos.y * CellSize);
+		sf::RectangleShape rectangle(sf::Vector2f(static_cast<float>(obj.transform.scale.x * CellSize), static_cast<float>(obj.transform.scale.y * CellSize)));
+		rectangle.setPosition(static_cast<float>(obj.transform.pos.x * CellSize), static_cast<float>(obj.transform.pos.y * CellSize));
 		Window.draw(rectangle);
 	}
 
@@ -34,10 +34,10 @@ namespace
 		{
 			uint64_t i = index * 4;
 
-			vertexArray[i + 0].position = sf::Vector2f(obj[i].GetVerticePosition<Side::TopLeft>().x, obj[i].GetVerticePosition<Side::TopLeft>().y);
-			vertexArray[i + 1].position = sf::Vector2f(obj[i].GetVerticePosition<Side::TopRight>().x, obj[i].GetVerticePosition<Side::TopRight>().y);
-			vertexArray[i + 2].position = sf::Vector2f(obj[i].GetVerticePosition<Side::BottomRight>().x, obj[i].GetVerticePosition<Side::BottomRight>().y);
-			vertexArray[i + 3].position = sf::Vector2f(obj[i].GetVerticePosition<Side::BottomLeft>().x, obj[i].GetVerticePosition<Side::BottomLeft>().y);
+			vertexArray[i + 0].position = sf::Vector2f((float)obj[i].GetVerticePosition<Side::TopLeft>().x, (float)obj[i].GetVerticePosition<Side::TopLeft>().y);
+			vertexArray[i + 1].position = sf::Vector2f((float)obj[i].GetVerticePosition<Side::TopRight>().x, (float)obj[i].GetVerticePosition<Side::TopRight>().y);
+			vertexArray[i + 2].position = sf::Vector2f((float)obj[i].GetVerticePosition<Side::BottomRight>().x, (float)obj[i].GetVerticePosition<Side::BottomRight>().y);
+			vertexArray[i + 3].position = sf::Vector2f((float)obj[i].GetVerticePosition<Side::BottomLeft>().x, (float)obj[i].GetVerticePosition<Side::BottomLeft>().y);
 
 			vertexArray[i + 0].color = sf::Color(obj[i].GetVertice<Side::TopLeft>().color.r, obj[i].GetVertice<Side::TopLeft>().color.g, obj[i].GetVertice<Side::TopLeft>().color.b, obj[i].GetVertice<Side::TopLeft>().color.a);
 			vertexArray[i + 1].color = sf::Color(obj[i].GetVertice<Side::TopRight>().color.r, obj[i].GetVertice<Side::TopLeft>().color.g, obj[i].GetVertice<Side::TopLeft>().color.b, obj[i].GetVertice<Side::TopLeft>().color.a);
